Remove Singleton from FoundationsManager, split FoundationsManager
